+++ components/pytes_e_box/text_sensor/pytes_e_box_battery_cell_text_sensor.h
@@ -8,7 +8,7 @@ namespace pytes_e_box {
 
 class PytesEBoxBatteryCellTextSensor : public PytesEBoxListener, public Component {
  public:
-  PytesEBoxBatteryCellTextSensor(int8_t bat_num,int8_t cell_num); 
+  PytesEBoxBatteryCellTextSensor(int bat_num,int cell_num); 
 
   void dump_config() override;
 
@@ -22,8 +22,8 @@ class PytesEBoxBatteryCellTextSensor : public PytesEBoxListener, public Componen
   virtual void on_batn_line_read(bat_index_LineContents *line);
 
  protected:
-  int8_t bat_num_;
-  int8_t cell_num_;
+  int  bat_num_;
+  int  cell_num_;
 };
 
 }  // namespace pytes_e_box
